got persistence just about worked out

+++ lib/blargh/post/persistence.rb
@@ -1,6 +1,13 @@
 module Blargh
   class Post
-    class ValidationError < Exception
+    def self.create(attributes = {})
+      new(attributes).tap(&:save)
+    end
+
+    def self.create!(attributes = {})
+      post = new(attributes)
+      raise ValidationError if !post.save
+      post
     end
 
     def new_record?
@@ -14,7 +21,15 @@ module Blargh
     def save
       @previously_changed = changes
       @changed_attributes.clear
-      valid? ? save_to_file : false
+      valid? ? create_or_update_file : false
+    end
+
+    def create_or_update_file
+      new_record? ? create_file : update_file
+    end
+
+    def update_file
+      true
     end
 
     def save!
@@ -23,45 +38,86 @@ module Blargh
     end
 
     # TODO: make private
-    def save_to_file
-      # TODO: check validations on creation if new record
+    def create_file
+      return false if !valid? || persisted? # don't create if we exist or are invalid
       @new_record = false
-      FileUtils.mkdir_p(directory)
+      FileUtils.mkdir_p(Post.directory)
+      file = save_to
       File.open(file, 'w') {|f| f.write(to_file) }
+      @saved_to = file
+      true
     end
 
     def has_file?
-      File.exists?(file)
+      @saved_to ? File.exists?(@saved_to) : false
+    end
+
+    # a new post's proposed home
+    def save_to
+      time_stamp = lambda { Time.now.to_i }.call
+      @attributes['id'] = get_unique_id(time_stamp)
+
+      "#{ Post.directory }/#{ read_attribute(:id) }-#{ slug }.#{extension}"
+    end
+
+    # good enough for now
+    def get_unique_id(proposed)
+      proposed = proposed.to_s
+      @tries ||= 0
+      @tries += 1
+
+      ids = Dir["#{ Post.directory }/*.textile"].map do |file|
+        if File.basename(file) =~ /\A(\d*)-/
+          Regexp.last_match(1)
+        end
+      end
+
+      if ids.include?(proposed)
+        get_unique_id("#{ proposed }#{ @tries }")
+      else
+        proposed
+      end
     end
 
     # TODO: make this configurable
     def basename
-      stamp = Time.now.to_i
-      "#{ stamp }-#{ slug }"
+      stamp = lambda { Time.now.to_i }
+      # "#{ id }-#{ slug }.#{extension}"
+      "#{ stamp.call }-#{ slug }.#{extension}"
     end
 
     def extension
-      '.textile'
+      'textile'
     end
 
     def file
-      "#{ directory }/#{ basename }#{ extension }"
+      "#{ Post.directory }/#{ basename }"
     end
 
-    def directory
-      # TODO: move this to an initialization of some sort
-      raise 'You need to configure your Blargh!' if Blargh.root.nil?
-      "#{ Blargh.root }/#{ Blargh.posts_directory }"
-    end
+    class << self
 
-    # converts our object into a file, pretty neat
-    def to_file
-      "farts"
-    end
+      def directory
+        # TODO: move this to an initialization of some sort
+        raise 'You need to configure your Blargh!' if Blargh.root.nil?
+        "#{ Blargh.root }/#{ Blargh.posts_directory }"
+      end
 
-    # TODO: come back to this to get the tags working
-    def self.reflect_on_association(association)
-      nil
+      # TODO: come back to this to get the tags working
+      def reflect_on_association(association)
+        nil
+      end
+
+      def delete_all
+        hit_list = Dir["#{ directory }/*"]
+        body_count = hit_list.count
+        FileUtils.rm_r(hit_list, :secure => true)
+        body_count
+      end
+      alias :destroy_all :delete_all
+
+      def count
+        Dir["#{ directory }/*"].count
+      end
     end
   end
 end
